backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
psycopg[binary]>=3.1
celery>=5.4
redis>=5.0
